Removed repeated code
