+++ lib/src/components/fields/yg_dropdown_field/yg_drop_down_child_route_mixin.dart
@@ -1,0 +1,35 @@
+import 'package:flutter/material.dart';
+import 'package:yggdrasil/src/components/fields/_fields.dart';
+
+mixin YgDropDownChildRouteMixin<T extends Object> on PopupRoute<Widget> {
+  YgAnyDropdownController<T> get dropdownController;
+  VoidCallback get onClose;
+
+  @override
+  void onPopInvoked(bool didPop) {
+    if (!didPop || !dropdownController.attached) {
+      return;
+    }
+
+    onClose();
+    dropdownController.discardChanges();
+  }
+
+  @override
+  void install() {
+    dropdownController.addListener(_handleDropdownControllerChange);
+    super.install();
+  }
+
+  @override
+  void dispose() {
+    dropdownController.removeListener(_handleDropdownControllerChange);
+    super.dispose();
+  }
+
+  void _handleDropdownControllerChange() {
+    if (!dropdownController.attached) {
+      navigator?.removeRoute(this);
+    }
+  }
+}
